feat(nodeSetup): Creating a node env
