+++ UI/entries.html
@@ -1,0 +1,77 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title></title>
+    <link rel="stylesheet" href="./styles/style.css">
+    <link href="https://stackpath.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet" integrity="sha384-wvfXpqpZZVQGK6TAh5PVlGOfQNHSoD2xbE+QkPxCAFlNEevoEH3Sl0sibVcOQVnN" crossorigin="anonymous">
+  </head>
+  <body class="body">
+      <header>
+          <ul>
+              <li><a href="./index.html">SignIn or SignUp</a></li>
+              <li><a class="active" href="#">Entries</a></li>
+              <li><a href="./profile.html">Profile</a></li>
+          </ul>
+      </header>
+    <table id="entries">
+      <tr>
+        <th>Id</th>
+        <th>Content</th>
+        <th>Action</th>
+      </tr>
+      <tr>
+        <td>1</td>
+        <td>Maria Anders hfhss hryjjf bdshdsh hjdsjsjd hsdfsdksdksdyty5</td>
+        <td><a class="modify" href="./view.html">View</a><a class="modify" href="./modifyEntry.html">Modify</a><a class="modify" href="#">Delete</a></td>
+      </tr>
+      <tr>
+        <td>2</td>
+        <td>Christina Berglund yhukiuytrefvhyk hds jyufyufyfrt hgvu</td>
+        <td><a class="modify" href="./view.html">View</a><a class="modify" href="./modifyEntry.html">Modify</a><a class="modify" href="#">Delete</a></td>
+      </tr>
+      <tr>
+        <td>3</td>
+        <td>Francisco Chang uiuyuyyi8ythfc bjknkln gjlngtgf jgjhytfth ykghh </td>
+        <td><a class="modify" href="./view.html">View</a><a class="modify" href="./modifyEntry.html">Modify</a><a class="modify" href="#">Delete</a></td>
+      </tr>
+      <tr>
+        <td>4</td>
+        <td>Roland Mendel bndufjsbuobvu sdhfiuaei udfvbgkusdaiubirsk dsbi</td>
+        <td><a class="modify" href="./view.html">View</a><a class="modify" href="./modifyEntry.html">Modify</a><a class="modify" href="#">Delete</a></td>
+      </tr>
+      <tr>
+        <td>5</td>
+        <td>Helen Bennett  bndufjsbuobvu sdhfiuaei udfvbgkusdaiubirsk dsbi</td>
+        <td><a class="modify" href="./view.html">View</a><a class="modify" href="./modifyEntry.html">Modify</a><a class="modify" href="#">Delete</a></td>
+      </tr>
+      <tr>
+        <td>6</td>
+        <td>Philip Cramer  bndufjsbuobvu sdhfiuaei udfvbgkusdaiubirsk dsbi</td>
+        <td><a class="modify" href="./view.html">View</a><a class="modify" href="./modifyEntry.html">Modify</a><a class="modify" href="#">Delete</a></td>
+      </tr>
+      <tr>
+        <td>7</td>
+        <td>Yoshi Tannamuri  bndufjsbuobvu sdhfiuaei udfvbgkusdaiubirsk dsbi</td>
+        <td><a class="modify" href="./view.html">View</a><a class="modify" href="./modifyEntry.html">Modify</a><a class="modify" href="#">Delete</a></td>
+      </tr>
+      <tr>
+        <td>9</td>
+        <td>Giovanni Rovelli  bndufjsbuobvu sdhfiuaei udfvbgkusdaiubirsk dsbi</td>
+        <td><a class="modify" href="./view.html">View</a><a class="modify" href="./modifyEntry.html">Modify</a><a class="modify" href="#">Delete</a></td>
+      </tr>
+      <tr>
+        <td>10</td>
+        <td>Simon Crowther  bndufjsbuobvu sdhfiuaei udfvbgkusdaiubirsk dsbi</td>
+        <td><a class="modify" href="./view.html">View</a><a class="modify" href="./modifyEntry.html">Modify</a><a class="modify" href="#">Delete</a></td>
+      </tr>
+      <tr>
+        <td>11</td>
+        <td>Marie Bertrand  bndufjsbuobvu sdhfiuaei udfvbgkusdaiubirsk dsbi</td>
+        <td><a class="modify" href="./view.html">View</a><a class="modify" href="./modifyEntry.html">Modify</a><a class="modify" href="#">Delete</a></td>
+      </tr>
+    </table>
+    <hr/>
+    <a href="./newEntries.html" class="modify">Add Entry</a>
+</html>
