+++ UI/index.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title></title>
+    <link rel="stylesheet" href="./styles/style.css">
+  </head>
+  <body>
+    <div class="divide left">
+        <header>
+            <ul>
+                <li><a class="active" href="#">SignIn or SignUp</a></li>
+                <li><a href="./entries.html">Entries</a></li>
+                <li><a href="./profile.html">Profile</a></li>
+            </ul>
+        </header>
+            <form>
+              <div id="center">
+                <h1>Login</h1>
+              <input type="email" name="email" placeholder="email">
+              <input type="password" name="password" placeholder="password">
+              <input type="submit" value="Login" class="button">
+            </div>
+            </form>
+    </div>
+
+    <div class="divide right">
+            <div class="container">
+                    <p> <h1> Dont have an Account</h1><b><a href="signup.html">Click here to Register</a></b></p>
+            </div>
+    </div>
+  </body>
+</html>
