--- modifyEntry.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title></title>
-    <link rel="stylesheet" href="./styles/style.css">
-  </head>
-  <body class="body">
-      <header>
-          <ul>
-              <li><a href="./index.html">SignIn or SignUp</a></li>
-              <li><a href="./entries.html">Entries</a></li>
-              <li><a href="./profile.html">Profile</a></li>
-          </ul>
-      </header>
-    <form class="bg-white">
-      <div id="center">
-        <h1>Modify Entry</h1>
-      <input type="text" name="modifyEntry" placeholder="Add Entry">
-      <textarea placeholder="Content"></textarea><br>
-      <input type="submit" value="Update Entry" class="button-add">
-      <input type="submit" value="Cancel Changes" class="button-add">
-    </div>
-    </form>
-  </body>
-</html>
