+++ package.json
@@ -1,0 +1,33 @@
+{
+  "name": "mydiary",
+  "version": "1.0.0",
+  "description": "An online journal where users can pen down their thoughts and feelings.",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "set DEBUG=index && nodemon server/index.js",
+    "lint": "eslint server/index.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ktrixs/MyDiary.git"
+  },
+  "author": "Ktrixs",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/ktrixs/MyDiary/issues"
+  },
+  "homepage": "https://github.com/ktrixs/MyDiary#readme",
+  "dependencies": {
+    "express": "^4.17.1"
+  },
+  "devDependencies": {
+    "chalk": "^2.4.2",
+    "debug": "^4.1.1",
+    "eslint": "^6.5.1",
+    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-plugin-import": "^2.18.2",
+    "morgan": "^1.9.1",
+    "nodemon": "^1.19.4"
+  }
+}
