--- profile.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title></title>
-    <link rel="stylesheet" href="./styles/style.css">
-  </head>
-  <body class="body">
-      <header>
-          <ul>
-              <li><a href="./index.html">SignIn or SignUp</a></li>
-              <li><a href="./entries.html">Entries</a></li>
-              <li><a href="./profile.html">Profile</a></li>
-          </ul>
-      </header>
-      <div id="center" class="bg-white">
-        <img src="./images/profile.jpg" alt="" class="user-pic" />
-        <h2>Kayitare Thierry</h2>
-        <h4>__ entries created since Joining</h4>
-        <b>Notify Me: </b>
-        <input type="checkbox" id="reminder" hidden/>
-        <label class="switch" for="reminder"></label>
-      </div>
-  </body>
-</html>
