--- view.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title></title>
-    <link rel="stylesheet" href="./styles/style.css">
-  </head>
-  <body class="body">
-      <header>
-          <ul>
-              <li><a href="./index.html">SignIn or SignUp</a></li>
-              <li><a href="./entries.html">Entries</a></li>
-              <li><a href="./profile.html">Profile</a></li>
-          </ul>
-      </header>
-      <div id="center" class="bg-white">
-        <h1>Modify Entry</h1>
-        <p>Antecipando em um ano aquilo que estava previst, 
-                aça um cenário base para os próximos a, 
-                aça um cenário base para os próximos a
-        </p>
-        <p>Antecipando em um ano aquilo que estava previst, 
-                aça um cenário base para os próximos a, 
-                aça um cenário base para os próximos a
-        </p>
-        <p>Antecipando em um ano aquilo que estava previst, 
-                aça um cenário base para os próximos a, 
-                aça um cenário base para os próximos a
-        </p>
-        <p>Antecipando em um ano aquilo que estava previst, 
-                aça um cenário base para os próximos a, 
-                aça um cenário base para os próximos a
-        </p>
-        <p>Antecipando em um ano aquilo que estava previst, 
-                aça um cenário base para os próximos a, 
-                aça um cenário base para os próximos a
-        </p>
-        <a href="./entries.html" class="modify"> <<< Back</a> <a class="modify" href="./modifyEntry.html">Modify Entry</a>
-    </div>
-  </body>
-</html>
